+++ FirebaseAuth/Sources/SystemService/FIRAuthStoredUserManager.m
@@ -72,6 +72,7 @@ static NSString *kStoredUserCoderKey = @"firebase_auth_stored_user_coder_key";
 #pragma mark - User for Access Group
 
 - (FIRUser *)getStoredUserForAccessGroup:(NSString *)accessGroup
+             shareAuthStateAcrossDevices:(BOOL)shareAuthStateAcrossDevices
                        projectIdentifier:(NSString *)projectIdentifier
                                    error:(NSError *_Nullable *_Nullable)outError {
   NSMutableDictionary *query = [[NSMutableDictionary alloc] init];
@@ -80,6 +81,9 @@ static NSString *kStoredUserCoderKey = @"firebase_auth_stored_user_coder_key";
   query[(__bridge id)kSecAttrAccessGroup] = accessGroup;
   query[(__bridge id)kSecAttrService] = projectIdentifier;
   query[(__bridge id)kSecAttrAccount] = kSharedKeychainAccountValue;
+  if (shareAuthStateAcrossDevices) {
+    query[(__bridge id)kSecAttrSynchronizable] = (__bridge id)kCFBooleanTrue;
+  }
 
   NSData *data = [self.keychainServices getItemWithQuery:query error:outError];
   // If there's an outError parameter and it's populated, or there's no data, return.
@@ -105,17 +109,25 @@ static NSString *kStoredUserCoderKey = @"firebase_auth_stored_user_coder_key";
 }
 
 - (BOOL)setStoredUser:(FIRUser *)user
-       forAccessGroup:(NSString *)accessGroup
-    projectIdentifier:(NSString *)projectIdentifier
-                error:(NSError *_Nullable *_Nullable)outError {
+                 forAccessGroup:(NSString *)accessGroup
+    shareAuthStateAcrossDevices:(BOOL)shareAuthStateAcrossDevices
+              projectIdentifier:(NSString *)projectIdentifier
+                          error:(NSError *_Nullable *_Nullable)outError {
   NSMutableDictionary *query = [[NSMutableDictionary alloc] init];
   query[(__bridge id)kSecClass] = (__bridge id)kSecClassGenericPassword;
-  query[(__bridge id)kSecAttrAccessible] =
-      (__bridge id)kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly;
+  if (shareAuthStateAcrossDevices) {
+    query[(__bridge id)kSecAttrAccessible] = (__bridge id)kSecAttrAccessibleAfterFirstUnlock;
+  } else {
+    query[(__bridge id)kSecAttrAccessible] =
+        (__bridge id)kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly;
+  }
 
   query[(__bridge id)kSecAttrAccessGroup] = accessGroup;
   query[(__bridge id)kSecAttrService] = projectIdentifier;
   query[(__bridge id)kSecAttrAccount] = kSharedKeychainAccountValue;
+  if (shareAuthStateAcrossDevices) {
+    query[(__bridge id)kSecAttrSynchronizable] = (__bridge id)kCFBooleanTrue;
+  }
 
 #if TARGET_OS_WATCH
   NSKeyedArchiver *archiver = [[NSKeyedArchiver alloc] initRequiringSecureCoding:false];
@@ -138,12 +150,20 @@ static NSString *kStoredUserCoderKey = @"firebase_auth_stored_user_coder_key";
 }
 
 - (BOOL)removeStoredUserForAccessGroup:(NSString *)accessGroup
+           shareAuthStateAcrossDevices:(BOOL)shareAuthStateAcrossDevices
                      projectIdentifier:(NSString *)projectIdentifier
                                  error:(NSError *_Nullable *_Nullable)outError {
   NSMutableDictionary *query = [[NSMutableDictionary alloc] init];
   query[(__bridge id)kSecClass] = (__bridge id)kSecClassGenericPassword;
-  query[(__bridge id)kSecAttrAccessible] =
-      (__bridge id)kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly;
+  if (shareAuthStateAcrossDevices) {
+    query[(__bridge id)kSecAttrAccessible] = (__bridge id)kSecAttrAccessibleAfterFirstUnlock;
+  } else {
+    query[(__bridge id)kSecAttrAccessible] =
+        (__bridge id)kSecAttrAccessibleAfterFirstUnlockThisDeviceOnly;
+  }
+  if (shareAuthStateAcrossDevices) {
+    query[(__bridge id)kSecAttrSynchronizable] = (__bridge id)kCFBooleanTrue;
+  }
 
   query[(__bridge id)kSecAttrAccessGroup] = accessGroup;
   query[(__bridge id)kSecAttrService] = projectIdentifier;