+++ FirebaseDatabase/Sources/Api/Private/FTypedefs_Private.h
@@ -46,6 +46,8 @@ typedef BOOL (^fbt_bool_nsstring_node)(NSString *, id<FNode>);
 typedef void (^fbt_void_path_node_marray)(FPath *, id<FNode>, NSMutableArray *);
 typedef BOOL (^fbt_bool_void)(void);
 typedef void (^fbt_void_nsstring_nsstring)(NSString *str1, NSString *str2);
+typedef void (^fbt_void_nsstring_id_nsstring)(NSString *str1, id dict1,
+                                              NSString *str2);
 typedef void (^fbt_void_nsstring_nserror)(NSString *str, NSError *error);
 typedef BOOL (^fbt_bool_path)(FPath *str);
 typedef void (^fbt_void_id)(id data);