+++ Firestore/Protos/nanopb/google/firestore/v1/write.nanopb.cc
@@ -37,13 +37,14 @@ using nanopb::PrintTail;
 
 
 
-const pb_field_t google_firestore_v1_Write_fields[7] = {
+const pb_field_t google_firestore_v1_Write_fields[8] = {
     PB_ANONYMOUS_ONEOF_FIELD(operation,   1, MESSAGE , ONEOF, STATIC  , FIRST, google_firestore_v1_Write, update, update, &google_firestore_v1_Document_fields),
     PB_ANONYMOUS_ONEOF_FIELD(operation,   2, BYTES   , ONEOF, POINTER , UNION, google_firestore_v1_Write, delete_, delete_, 0),
     PB_ANONYMOUS_ONEOF_FIELD(operation,   5, BYTES   , ONEOF, POINTER , UNION, google_firestore_v1_Write, verify, verify, 0),
     PB_ANONYMOUS_ONEOF_FIELD(operation,   6, MESSAGE , ONEOF, STATIC  , UNION, google_firestore_v1_Write, transform, transform, &google_firestore_v1_DocumentTransform_fields),
     PB_FIELD(  3, MESSAGE , OPTIONAL, STATIC  , OTHER, google_firestore_v1_Write, update_mask, transform, &google_firestore_v1_DocumentMask_fields),
     PB_FIELD(  4, MESSAGE , OPTIONAL, STATIC  , OTHER, google_firestore_v1_Write, current_document, update_mask, &google_firestore_v1_Precondition_fields),
+    PB_FIELD(  7, MESSAGE , REPEATED, POINTER , OTHER, google_firestore_v1_Write, update_transforms, current_document, &google_firestore_v1_DocumentTransform_FieldTransform_fields),
     PB_LAST_FIELD
 };
 
@@ -159,6 +160,10 @@ std::string google_firestore_v1_Write::ToString(int indent) const {
     if (has_current_document) {
         result += PrintMessageField("current_document ",
             current_document, indent + 1, true);
+    }
+    for (pb_size_t i = 0; i != update_transforms_count; ++i) {
+        result += PrintMessageField("update_transforms ",
+            update_transforms[i], indent + 1, true);
     }
 
     std::string tail = PrintTail(indent);